requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
numpy>=1.26.0
pandas>=2.1.0
Pillow>=10.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0
nbformat>=5.9.0
jupyter>=1.0.0
